Checking in everything I merged down last week.
git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C394

+++ WebAPI/API/C_Development/Work/Winston/Emar.Core/Orders/Model/DepartmentPreferredItemDto.cs
@@ -8,10 +8,9 @@ namespace Emar.Core.Orders.Model
         /// <summary>
         /// User Quick Lists are site-specific
         /// </summary>
-        public long SiteId { get; set; }
+        public int SiteId { get; set; }
 
         public string DepartmentCode { get; set; }
-
 
         public IEnumerable<HateOasLinkDto> Links;
     }
