+++ WebAPI/API/C_Development/Work/Winston/Emar.Core/Orders/Model/UserQuickListItemDto.cs
@@ -13,7 +13,7 @@ namespace Emar.Core.Orders.Model
         /// <summary>
         /// User Quick Lists are site-specific
         /// </summary>
-        public long SiteId { get; set; }
+        public int SiteId { get; set; }
 
         public IEnumerable<HateOasLinkDto> Links;
     }
